test: relocate provider login and refresh cases to agent/provider-runtime.test.ts
Second commit of PR 1. The first commit moved the code with the test files
untouched, which is the evidence the refactor preserved behaviour; this one
moves the tests to sit next to the module they now exercise.

Relocation, not a rewrite: every assertion is byte-identical, each test still
constructs a real AgentService through agent-service-test-harness.ts and
asserts through the public status/event surface. Seven declarations (eight
cases) move — the concurrent account check with the curated Codex model list,
the ChatGPT App Server login, the it.each over the bundled Claude/Grok login
commands, the failed login page, the login timeout, independent logins under
Refresh, and the keep-the-active-client reconnect.

The it.each keeps its dynamic `service[connect]()` call with the
"connectClaude"/"connectGrok" literals, so both stay facade methods.

Everything left in agent-service.providers.test.ts crosses domains — turn
progress, duplication, file resolution, tags, thread creation, browser tool
routing — and stays in the shard named for the behaviour.

Tests: agent/provider-runtime.test.ts 8 passed, agent-service.providers.test.ts
14 passed; 22 together, the same 22 the shard ran before the split. tsc on
tsconfig.node.json and biome on both files clean.
