+++ src/backend/agent-service.providers.test.ts
@@ -12,9 +12,6 @@ import { AgentRoutineStore } from "./agent-routine-store";
 import { AgentService } from "./agent-service";
 import {
   CREATE_BOT_INPUT,
-  createFakeClaude,
-  createFakeGrok,
-  createPendingFakeClaude,
   EMPTY_LAYOUT,
   FakeAgentClient,
   fakeBrowser,
@@ -22,7 +19,6 @@ import {
   notification,
   paramsRecord,
   protocolMessages,
-  readTextOrEmpty,
   startAgentTestFixture,
   stopAgentTestFixture,
   stores,
@@ -110,68 +106,6 @@ describe.sequential("AgentService: providers", () => {
       notification("turn/completed", { threadId, turn: { id: turnId, status: "completed" } }),
     );
     await waitFor(() => events.some((event) => event.type === "turn-completed" && event.turnId === turnId));
-  });
-
-  it("checks providers concurrently and publishes each completed row", async () => {
-    process.env.OPENBOT_CLAUDE_PATH = await createFakeClaude(root);
-    process.env.OPENBOT_GROK_PATH = await createFakeGrok(root);
-    const { store, mailbox } = stores(root);
-    const delays: Record<AgentProvider, number> = { codex: 60, claude: 5, grok: 30 };
-    const availableOrder: AgentProvider[] = [];
-    const seen = new Set<AgentProvider>();
-    const accountReads = new Set<AgentProvider>();
-    let releaseAccountReads: (() => void) | undefined;
-    const allAccountReadsStarted = new Promise<void>((resolve) => {
-      releaseAccountReads = resolve;
-    });
-    const waitForConcurrentAccountReads = async (method: string, provider: AgentProvider) => {
-      if (method !== "account/read") return;
-      accountReads.add(provider);
-      if (accountReads.size === 3) releaseAccountReads?.();
-      await allAccountReadsStarted;
-    };
-    service = new AgentService(
-      store,
-      mailbox,
-      fakeBrowser(),
-      30_000,
-      "codex",
-      (provider) =>
-        new FakeAgentClient(
-          provider,
-          "DONE",
-          true,
-          true,
-          { "account/read": delays[provider] },
-          waitForConcurrentAccountReads,
-        ),
-    );
-    service.on("event", (event) => {
-      if (event.type !== "status") return;
-      for (const provider of event.status.providers ?? []) {
-        if (provider.state !== "available" || seen.has(provider.id)) continue;
-        seen.add(provider.id);
-        availableOrder.push(provider.id);
-      }
-    });
-
-    await Promise.race([
-      service.initialize(),
-      new Promise<never>((_, reject) =>
-        setTimeout(() => reject(new Error("Provider account checks did not start concurrently.")), 3_000),
-      ),
-    ]);
-
-    expect(availableOrder).toEqual(["claude", "grok", "codex"]);
-
-    // The fake advertises gpt-5.5, gpt-5.4, gpt-5.4-mini and gpt-5.3-codex-spark
-    // alongside the curated three, so CURATED_CODEX_MODEL_IDS has to drop four.
-    expect(
-      service
-        .listModels()
-        .filter((model) => model.provider === "codex")
-        .map((model) => model.id),
-    ).toEqual(["gpt-5.6-luna", "gpt-5.6-terra", "gpt-5.6-sol"]);
   });
 
   it("duplicates persistent agent data without conversation or routine-run history", async () => {
@@ -625,242 +559,6 @@ describe.sequential("AgentService: providers", () => {
       expect(params.runtimeWorkspaceRoots).toEqual([params.cwd, store.sharedRoot]);
     }
     expect((await store.getOrCreate("chief")).threadId).not.toBe((await store.getOrCreate("sales-outbound")).threadId);
-  });
-
-  it("connects ChatGPT through the Codex App Server and promotes the authenticated client", async () => {
-    const { store, mailbox } = stores(root);
-    const codexClients: FakeAgentClient[] = [];
-    const openExternal = vi.fn(async () => undefined);
-    service = new AgentService(store, mailbox, fakeBrowser(), 30_000, "codex", (provider) => {
-      const client = new FakeAgentClient(
-        provider,
-        provider === "codex" ? "CODEX_DONE" : "CLAUDE_DONE",
-        true,
-        provider !== "codex",
-      );
-      if (provider === "codex") codexClients.push(client);
-      return client;
-    });
-    await service.initialize();
-
-    expect(service.getStatus().providers).toContainEqual(
-      expect.objectContaining({ id: "codex", state: "sign-in-required" }),
-    );
-    const connecting = await service.connectChatGPT(openExternal);
-
-    expect(connecting.providers).toContainEqual(
-      expect.objectContaining({
-        id: "codex",
-        state: "sign-in-required",
-        connectionState: "connecting",
-        version: "0.144.1",
-      }),
-    );
-    expect(openExternal).toHaveBeenCalledWith("https://auth.openai.test/connect");
-    expect(codexClients).toHaveLength(2);
-    expect(codexClients[1]?.requests).toContainEqual({
-      method: "account/login/start",
-      params: {
-        type: "chatgpt",
-        appBrand: "chatgpt",
-        codexStreamlinedLogin: true,
-        useHostedLoginSuccessPage: true,
-      },
-    });
-
-    await service.connectChatGPT(openExternal);
-    expect(openExternal).toHaveBeenCalledTimes(2);
-    expect(codexClients).toHaveLength(3);
-    expect(codexClients[1]?.requests).toContainEqual({
-      method: "account/login/cancel",
-      params: { loginId: "login-1" },
-    });
-    expect(codexClients[1]?.running).toBe(false);
-    codexClients[1]?.completeLogin(true);
-    expect(service.getStatus().providers).toContainEqual(
-      expect.objectContaining({ id: "codex", connectionState: "connecting" }),
-    );
-    codexClients[2]?.completeLogin(true);
-    await waitFor(
-      () => service?.getStatus().providers?.find((provider) => provider.id === "codex")?.state === "available",
-    );
-
-    expect(service.getStatus().phase).toBe("ready");
-    expect(service.getStatus().providers).toContainEqual(
-      expect.objectContaining({ id: "codex", state: "available", email: "codex@example.com" }),
-    );
-  });
-
-  it.each([
-    { target: "claude", connect: "connectClaude", pathVariable: "OPENBOT_CLAUDE_PATH", createCli: createFakeClaude },
-    { target: "grok", connect: "connectGrok", pathVariable: "OPENBOT_GROK_PATH", createCli: createFakeGrok },
-  ] as const)(
-    "connects $target through the bundled CLI login command",
-    async ({ target, connect, pathVariable, createCli }) => {
-      process.env[pathVariable] = await createCli(root);
-      const { store, mailbox } = stores(root);
-      let clients = 0;
-      service = new AgentService(store, mailbox, fakeBrowser(), 30_000, target, (provider) => {
-        const authenticated = provider === target ? clients > 0 : true;
-        if (provider === target) clients += 1;
-        return new FakeAgentClient(provider, "DONE", true, authenticated);
-      });
-      await service.initialize();
-
-      expect(service.getStatus().providers).toContainEqual(
-        expect.objectContaining({ id: target, state: "sign-in-required" }),
-      );
-
-      const connecting = await service[connect]();
-
-      expect(connecting.providers).toContainEqual(
-        expect.objectContaining({ id: target, state: "sign-in-required", connectionState: "connecting" }),
-      );
-      await waitFor(() => clients === 2);
-      await waitFor(
-        () => service?.getStatus().providers?.find((provider) => provider.id === target)?.state === "available",
-      );
-      expect(service.getStatus().providers).toContainEqual(
-        expect.objectContaining({ id: target, state: "available", email: `${target}@example.com` }),
-      );
-    },
-  );
-
-  it("restores the connect action when the login page cannot open", async () => {
-    const { store, mailbox } = stores(root);
-    service = new AgentService(
-      store,
-      mailbox,
-      fakeBrowser(),
-      30_000,
-      "codex",
-      (provider) => new FakeAgentClient(provider, "DONE", true, provider !== "codex"),
-    );
-    await service.initialize();
-
-    await expect(service.connectChatGPT(async () => Promise.reject(new Error("browser failed")))).rejects.toThrow(
-      "could not open",
-    );
-    expect(service.getStatus().providers).toContainEqual(
-      expect.objectContaining({ id: "codex", state: "sign-in-required" }),
-    );
-  });
-
-  it("cancels a ChatGPT login that does not complete", async () => {
-    const { store, mailbox } = stores(root);
-    const codexClients: FakeAgentClient[] = [];
-    service = new AgentService(store, mailbox, fakeBrowser(), 30_000, "codex", (provider) => {
-      const client = new FakeAgentClient(provider, "DONE", true, provider !== "codex");
-      if (provider === "codex") codexClients.push(client);
-      return client;
-    });
-    await service.initialize();
-    vi.useFakeTimers();
-    await service.connectChatGPT(async () => undefined);
-
-    await vi.advanceTimersByTimeAsync(10 * 60_000);
-
-    expect(codexClients[1]?.requests).toContainEqual({
-      method: "account/login/cancel",
-      params: { loginId: "login-1" },
-    });
-    expect(service.getStatus().providers).toContainEqual(
-      expect.objectContaining({
-        id: "codex",
-        state: "sign-in-required",
-        message: expect.stringContaining("timed out"),
-      }),
-    );
-  });
-
-  it("runs provider logins independently and Refresh cancels both generations", async () => {
-    const claudeLoginLog = join(root, "claude-login.log");
-    process.env.OPENBOT_FAKE_CLAUDE_LOGIN_LOG = claudeLoginLog;
-    process.env.OPENBOT_CLAUDE_PATH = await createPendingFakeClaude(root);
-    const { store, mailbox } = stores(root);
-    const codexClients: FakeAgentClient[] = [];
-    service = new AgentService(store, mailbox, fakeBrowser(), 30_000, "codex", (provider) => {
-      const client = new FakeAgentClient(provider, "DONE", true, false);
-      if (provider === "codex") codexClients.push(client);
-      return client;
-    });
-    await service.initialize();
-
-    await Promise.all([service.connectChatGPT(async () => undefined), service.connectClaude()]);
-    expect(service.getStatus().providers).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({ id: "codex", connectionState: "connecting" }),
-        expect.objectContaining({ id: "claude", connectionState: "connecting" }),
-      ]),
-    );
-    await waitFor(async () => (await readTextOrEmpty(claudeLoginLog)).includes("started"));
-
-    await service.connectClaude();
-    await waitFor(async () => {
-      const log = await readTextOrEmpty(claudeLoginLog);
-      return log.match(/^started$/gmu)?.length === 2 && log.includes("stopped");
-    });
-    expect(service.getStatus().providers).toContainEqual(
-      expect.objectContaining({ id: "claude", connectionState: "connecting" }),
-    );
-
-    await service.refreshProviders();
-
-    expect(codexClients[1]?.requests).toContainEqual({
-      method: "account/login/cancel",
-      params: { loginId: "login-1" },
-    });
-    expect(codexClients[1]?.running).toBe(false);
-    await waitFor(async () => (await readTextOrEmpty(claudeLoginLog)).match(/^stopped$/gmu)?.length === 2);
-    expect(service.getStatus().providers).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({ id: "codex", state: "sign-in-required" }),
-        expect.objectContaining({ id: "claude", state: "sign-in-required" }),
-      ]),
-    );
-    expect(service.getStatus().providers?.some((provider) => provider.connectionState === "connecting")).toBe(false);
-
-    // The stale login completion is queued behind the codex connection command
-    // that `refreshProviders` runs, so awaiting the refresh proves the service
-    // processed it and still refused to sign the cancelled generation in.
-    codexClients[1]?.completeLogin(true);
-    await service.refreshProviders();
-    expect(service.getStatus().providers).toContainEqual(
-      expect.objectContaining({ id: "codex", state: "sign-in-required" }),
-    );
-  });
-
-  it("keeps the active ChatGPT client until reconnect succeeds", async () => {
-    const { store, mailbox } = stores(root);
-    const codexClients: FakeAgentClient[] = [];
-    service = new AgentService(store, mailbox, fakeBrowser(), 30_000, "codex", (provider) => {
-      const client = new FakeAgentClient(provider, "DONE", true, provider !== "codex" || codexClients.length === 0);
-      if (provider === "codex") codexClients.push(client);
-      return client;
-    });
-    await service.initialize();
-    const activeClient = codexClients[0];
-
-    await service.connectChatGPT(async () => undefined);
-    expect(service.getStatus().providers).toContainEqual(
-      expect.objectContaining({ id: "codex", state: "available", connectionState: "connecting" }),
-    );
-    codexClients[1]?.completeLogin(false);
-    await waitFor(() => !service?.getStatus().providers?.find((provider) => provider.id === "codex")?.connectionState);
-    expect(activeClient?.running).toBe(true);
-    expect(service.getStatus().providers).toContainEqual(
-      expect.objectContaining({ id: "codex", state: "available", message: expect.stringContaining("not completed") }),
-    );
-
-    await service.connectChatGPT(async () => undefined);
-    codexClients[2]?.completeLogin(true);
-    await waitFor(
-      () =>
-        service?.getStatus().providers?.find((provider) => provider.id === "codex")?.state === "available" &&
-        !service?.getStatus().providers?.find((provider) => provider.id === "codex")?.connectionState,
-    );
-    expect(activeClient?.running).toBe(false);
-    expect(codexClients[2]?.running).toBe(true);
   });
 
   it("maps provider browser tool calls to the stable OpenBot thread", async () => {
